Refactor token-related functions into a new module and update references
